Separate reader from reporter, simplify metrics, other review comments
